minor refactoring

+++ main.cxx
@@ -29,7 +29,7 @@ fvec4 sometest()
 	[[ maybe_unused]] auto third3 = first + first2;
 
 
-	[[ maybe_unused]] bool b1 = within_box(ivec2(3, 5), ivec2(4, 4), (int)iscal(2));
+	[[ maybe_unused]] bool b1 = within_box(ivec2(3, 5), ivec2(4, 4), iscal(2));
 
 
 	dsga::storage_wrapper<int, 4> sw{ 999, 9999, 99999, 999999 };
